Create Mukesh and his friends have set out on a vacation to Coorg. They have booked accommodation in a resort and the resort authorities organize Campfires every night as a part of their daily activities. Mukesh volunteered himself for an activity called the "Adjacent Stick Game" where sticks of different lengths will be placed in a line and Mukesh needs to remove a stick from each adjacent pair of sticks. He then has to form a bigger stick by combining all the remaining sticks.    Mukesh needs to know the smallest length of the bigger stick so formed and needs your help to compute the same. Given the number of sticks N and the lengths of each of the sticks, write a program to find the smallest length of the bigger stick that is formed.
Single File Programming Question
Problem Statement



Mukesh and his friends have set out on a vacation to Coorg. They have booked accommodation in a resort and the resort authorities organize Campfires every night as a part of their daily activities. Mukesh volunteered himself for an activity called the "Adjacent Stick Game" where sticks of different lengths will be placed in a line and Mukesh needs to remove a stick from each adjacent pair of sticks. He then has to form a bigger stick by combining all the remaining sticks.



Mukesh needs to know the smallest length of the bigger stick so formed and needs your help to compute the same. Given the number of sticks N and the lengths of each of the sticks, write a program to find the smallest length of the bigger stick that is formed.

Input format :
The first line contains a single integer, n (1 ≤ n ≤ 10^5), representing the number of sticks.

The second line contains n integers a_1, a_2, ..., a_n (1 ≤ a_i ≤ 10^3), representing the lengths of the sticks.

Output format :
The output is a single integer representing the maximum possible sum of the lengths of adjacent sticks.

Code constraints :
The value of N should be an even integer, 1 ≤ N ≤ 50.

The length of each stick should be an integer, 1 ≤ a_i ≤ 1000.

Sample test cases :
Input 1 :
4
2 1 3 1
Output 1 :
2
Input 2 :
4
4 2 3 5
Output 2 :
5